Fix overdisposing of singletons
